# api-docs.json
{"openapi":"3.0.1","info":{"title":"REST API's RESTful by TaskHub Management Premium with Java, Spring, Docker and Kubernetes","description":"REST API's RESTful by TaskHub Management Premium with Java, Spring, Docker and Kubernetes","termsOfService":"https://joaojunio-dev.vercel.app","license":{"name":"Apache 2.0","url":"https://joaojunio-dev.vercel.app"},"version":"1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/person":{"get":{"tags":["Person"],"summary":"Finds all people","description":"Finds all people","operationId":"findAll","responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PersonController"}}}}},"400":{"description":"Bad Request"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"404":{"description":"Not Found"},"500":{"description":"Internal Server Error"}}},"put":{"tags":["Person"],"summary":"Update one person","description":"Update one person","operationId":"update","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PersonDTO"}},"application/xml":{"schema":{"$ref":"#/components/schemas/PersonDTO"}},"application/yaml":{"schema":{"$ref":"#/components/schemas/PersonDTO"}}},"required":true},"responses":{"200":{"description":"Success","content":{"application/json":{}}},"400":{"description":"Bad Request"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"404":{"description":"Not Found"},"500":{"description":"Internal Server Error"}}},"post":{"tags":["Person"],"summary":"Create person","description":"Create person","operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PersonDTO"}},"application/xml":{"schema":{"$ref":"#/components/schemas/PersonDTO"}},"application/yaml":{"schema":{"$ref":"#/components/schemas/PersonDTO"}}},"required":true},"responses":{"200":{"description":"Success","content":{"application/json":{}}},"400":{"description":"Bad Request"},"401":{"description":"Unauthorized"},"404":{"description":"Not Found"},"500":{"description":"Internal Server Error"}}}},"/auth/refresh/{username}":{"put":{"tags":["Authentication EndPoint!"],"operationId":"refreshToken","parameters":[{"name":"username","in":"path","required":true,"schema":{"type":"string"}},{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/auth/signin":{"post":{"tags":["Authentication EndPoint!"],"operationId":"signIn","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountCredentialsDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/auth/createUser":{"post":{"tags":["Authentication EndPoint!"],"operationId":"create_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountCredentialsDTO"}},"application/xml":{"schema":{"$ref":"#/components/schemas/AccountCredentialsDTO"}},"application/yaml":{"schema":{"$ref":"#/components/schemas/AccountCredentialsDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountCredentialsDTO"}},"application/xml":{"schema":{"$ref":"#/components/schemas/AccountCredentialsDTO"}},"application/yaml":{"schema":{"$ref":"#/components/schemas/AccountCredentialsDTO"}}}}}}},"/person/{id}":{"get":{"tags":["Person"],"summary":"Find by one person","description":"Find by one person","operationId":"findById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Bad Request"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"404":{"description":"Not Found"},"500":{"description":"Internal Server Error"}}},"delete":{"tags":["Person"],"summary":"Delete one person","description":"Delete one person","operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Success","content":{"application/json":{}}},"400":{"description":"Bad Request"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"404":{"description":"Not Found"},"500":{"description":"Internal Server Error"}}}}},"components":{"schemas":{"Links":{"type":"object","additionalProperties":{"$ref":"#/components/schemas/Link"}},"PersonDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"firstName":{"type":"string"},"lastName":{"type":"string"},"birthDate":{"type":"string","format":"date"},"phone":{"type":"string"},"_links":{"$ref":"#/components/schemas/Links"}}},"AccountCredentialsDTO":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"},"fullname":{"type":"string"}}},"PersonController":{"type":"object"},"Link":{"type":"object","properties":{"href":{"type":"string"},"hreflang":{"type":"string"},"title":{"type":"string"},"type":{"type":"string"},"deprecation":{"type":"string"},"profile":{"type":"string"},"name":{"type":"string"},"templated":{"type":"boolean"}}}}}}